+++ src/cpu/cpu.js
@@ -27,7 +27,8 @@ class Cpu {
     this.target = null;
     this.codes = this.mapCodeNameToCode();
     this.lambdas = this.mapCodeToLambda();
-    this.stack_order = this.mapStackOrder();
+    this.stackOrder = this.mapStackOrder();
+    this.registerIds = this.mapRegisterIdentifier();
     this.code = [cpus.TFRWTOOB, cpus.READWLOW, cpus.READHIGH];
     this.PC.set(0xfffe);
   }
@@ -88,6 +89,9 @@ class Cpu {
     result['BITTEST'] = cpus.BITTEST;
     result['COMPLEMENT'] = cpus.COMPLEMENT;
     result['NEGATE'] = cpus.NEGATE;
+    result['EXCHANGE'] = cpus.EXCHANGE;
+    result['TRANSFER'] = cpus.TRANSFER;
+    result['SIGNEXTEND'] = cpus.SIGNEXTEND;
     return result;
   }
 
@@ -151,6 +155,9 @@ class Cpu {
     result[cpus.BITTEST] = this.read_and_bit_test;
     result[cpus.COMPLEMENT] = this.complement_byte;
     result[cpus.NEGATE] = this.negate_byte;
+    result[cpus.EXCHANGE] = this.exchange;
+    result[cpus.TRANSFER] = this.transfer;
+    result[cpus.SIGNEXTEND] = this.sign_extend;
     return result;
   }
 
@@ -172,6 +179,25 @@ class Cpu {
   };
 
   /**
+   * Create ordered list of register identifiers
+   * @returns {[]} register list
+   */
+  mapRegisterIdentifier() {
+    const result = [];
+    result[0] = 'D';
+    result[1] = 'X';
+    result[2] = 'Y';
+    result[3] = 'U';
+    result[4] = 'S';
+    result[5] = 'PC';
+    result[8] = 'A';
+    result[9] = 'B';
+    result[10] = 'CC';
+    result[11] = 'DP';
+    return result;
+  }
+
+  /**
    * Populate the execution stack with micro-codes.
    * translates string codes to cpu codes
    * @param {[]} instructionCode list
@@ -179,7 +205,12 @@ class Cpu {
   populateCodeStack(instructionCode) {
     this.code = [];
     for (let index = 0; index < instructionCode.length; index++) {
-      this.code.unshift(this.codes[instructionCode[index]]);
+      const operation = this.codes[instructionCode[index]];
+      if (typeof operation === 'undefined') {
+        throw new Error ("unknown operation " + instructionCode[index]);
+      } else {
+        this.code.unshift(this.codes[instructionCode[index]]);
+      }
     }
   }
 
@@ -262,6 +293,15 @@ class Cpu {
     if (this.code.length === 0) {
       this.clearInstruction();
     }
+  }
+
+  /**
+   * evaluate register id to find register name
+   * @param {number} id 4bit identifier of register
+   * @returns {string} register name
+   */
+  identifyRegister(id) {
+    return this.registerIds[id];
   }
 
   swap_internal_registers = () => {
@@ -595,8 +635,41 @@ class Cpu {
     this.object.set(this.alu1.negate(this.object.fetch()));
   }
 
+  sign_extend = () => {
+    if ((this.registers.get('B').fetch() & 0x80) !== 0) {
+      this.registers.get('A').set(0xff);
+      this.CC.negative(true);
+    } else {
+      this.registers.get('A').set(0x00);
+      this.CC.negative(false);
+    }
+    this.CC.zero(this.registers.get('D').fetch() === 0);
+  }
+
+  exchange = () => {
+    const id1 = (this.W.fetch() & 0xf0) >> 4;
+    const id2 = this.W.fetch() & 0x0f;
+    if ((id1 & 8) === (id2 & 8)) {
+      const register1 = this.identifyRegister(id1);
+      const register2 = this.identifyRegister(id2);
+      const temporary = this.registers.get(register1).fetch();
+      this.registers.get(register1).set(this.registers.get(register2).fetch());
+      this.registers.get(register2).set(temporary);
+    }
+  }
+
+  transfer = () => {
+    const id1 = (this.W.fetch() & 0xf0) >> 4;
+    const id2 = this.W.fetch() & 0x0f;
+    if ((id1 & 8) === (id2 & 8)) {
+      const register1 = this.identifyRegister(id1);
+      const register2 = this.identifyRegister(id2);
+      this.registers.get(register2).set(this.registers.get(register1).fetch());
+    }
+  }
+
   select_register = (stackMask) => {
-    let register = this.stack_order[stackMask];
+    let register = this.stackOrder[stackMask];
     if (register === 'US') {
       if (this.object.name === 'U') {
         register = 'S';