+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Virtual 6809 CHANGELOG #
 
+## Unreleased ##
+* Add EXG and TFR instructions
+* Add Sign Extend (SEX) instruction
+
 ## V0.2 - Feb 11 2021 ##
 * Add BIT, COM and NEG instructions
 * Add shift and rotate instructions